+++ cmd/nivis/completion.go
@@ -4,6 +4,8 @@
 package main
 
 import (
+	"context"
+
 	"github.com/spf13/cobra"
 )
 
@@ -17,7 +19,7 @@ func stateIDs(_ *cobra.Command, args []string, _ string) ([]string, cobra.ShellC
 	if len(args) > 0 {
 		return nil, cobra.ShellCompDirectiveNoFileComp
 	}
-	store, err := openStore()
+	store, err := openStore(context.Background())
 	if err != nil {
 		return nil, cobra.ShellCompDirectiveNoFileComp
 	}